requirement.txt
langchain
python-dotenv
langchain-openai
ipykernel
beautifulsoup4
langchain_community
faiss-cpu
streamlit

